+++ test/integration/abstract-sablier-v2/AbstractSablierV2.t.sol
@@ -22,7 +22,7 @@ contract AbstractSablierV2 is SablierV2 {
         return address(0);
     }
 
-    function getDepositAmount(uint256 streamId) external pure override returns (uint256) {
+    function getDepositAmount(uint256 streamId) external pure override returns (uint128) {
         streamId;
         return 0;
     }
@@ -32,7 +32,7 @@ contract AbstractSablierV2 is SablierV2 {
         return address(0);
     }
 
-    function getReturnableAmount(uint256 streamId) external pure override returns (uint256) {
+    function getReturnableAmount(uint256 streamId) external pure override returns (uint128) {
         streamId;
         return 0;
     }
@@ -42,22 +42,22 @@ contract AbstractSablierV2 is SablierV2 {
         return address(0);
     }
 
-    function getStartTime(uint256 streamId) external pure override returns (uint64) {
+    function getStartTime(uint256 streamId) external pure override returns (uint40) {
         streamId;
         return 0;
     }
 
-    function getStopTime(uint256 streamId) external pure override returns (uint64) {
+    function getStopTime(uint256 streamId) external pure override returns (uint40) {
         streamId;
         return 0;
     }
 
-    function getWithdrawableAmount(uint256 streamId) external pure override returns (uint256) {
+    function getWithdrawableAmount(uint256 streamId) external pure override returns (uint128) {
         streamId;
         return 0;
     }
 
-    function getWithdrawnAmount(uint256 streamId) external pure override returns (uint256) {
+    function getWithdrawnAmount(uint256 streamId) external pure override returns (uint128) {
         streamId;
         return 0;
     }
@@ -150,7 +150,7 @@ contract AbstractSablierV2 is SablierV2 {
     function _withdraw(
         uint256 streamId,
         address to,
-        uint256 amount
+        uint128 amount
     ) internal pure override {
         streamId;
         to;
